+++ source/Internal/Domain/Admin/Event/AdminModeChangedEvent.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\EshopCommunity\Internal\Domain\Admin\Event;
+
+use Symfony\Contracts\EventDispatcher\Event;
+
+class AdminModeChangedEvent extends Event
+{
+}